Compile warning fix.

+++ XMDS/XMDSAppDelegate.m
@@ -634,7 +634,7 @@
      Gestalt(gestaltSystemVersionMinor, &minor);
      Gestalt(gestaltSystemVersionBugFix, &bugfix);
      
-     return [NSString stringWithFormat:@"%d.%d.%d", major, minor, bugfix];
+     return [NSString stringWithFormat:@"%d.%d.%d", (int)major, (int)minor, (int)bugfix];
  }
